Only perform reconnect on send and use new connection
previously a reconnect attempt was made, but the connection was not
used. This meant that when a connection was lost the next 2 commands
would fail. The first because the read fails and the second because
the reconnect is needed.

Now the command after the disconnect fails, but the next command
will attempt a reconnect and use the connection