Use a stable machine identifier for OID generation
Instead of filling the host-specific identifier of OIDs with different
randomness for each generated id, just generate randomness once and use it as
the host part for all OIDs generated by that process and all threads it might
create.

It's still not machine-specific at all, but *much* closer to what the spec
expects us to do than what was there before.

It shall also be noted that, by doing this in perl space, possible issues due to
rand not being re-entrant and rand_r not being portable are avoided, and that
perl's rand will be properly seeded automatically, while the c version has some
global seed state that should be initialised but wasn't.