Allow enabling SSL/SASL builds via environment vars
When MongoDB is a deep dependency of other modules, a typical CPAN
install won't give an opportunity to enable SSL or SASL.

This patch allows enabling them via environment variables which
can be set as part of the overall application dependencies process.

It also fixes a bug using --ssl and --sasl together. Now they are
additive instead of the last one on the command line winning.  (I'm not
sure if it makes sense to enable both, but in that case, it should error
out, not just silently drop one the way it used to.)