PERL-318 fix compiler warnings from xs/BSON.xs
This change makes 'decode_bson' inflate regular expressions.  This is an
arbitrary choice, but consistent with how it inflates dbrefs.  No code
appears to use 'decode_bson' -- it seems to be a legacy function
supporting 'read_documents' in Mongo.pm -- so I think this arbitrary
choice is acceptable.

The other fix is a cast to the correct type for newSVpvn.