Avoid making an index direction look like a string
The code for determining if a scalar is a string or a number for
serialization gets tripped up if the scalar has been used as
a string. This cropped up in tests with Perl v5.8.5 so it may
not be an issue on v5.10+.

The commit copies the value in question before testing it to
preserve its original form.