Implement wire protocol in pure Perl
This is not portable prior to 5.10, so this is likely to be temporary
and we'll probably want to make pack templates variable based on
architecture.  E.g. little-endian is fine on 5.8, but big-endian will
require 5.10 for pack features to fix endian.

It's also not not portable to 32-bit systems, but we can fix that easily
later by making cursor IDs opaque once we control the end-to-end
query/reply process in pure Perl.