fix readpref.t to skip in some configurations
With find_master enabled in t/lib/MongoDBTest.pm, the readpref.t
tests fail if the default DB is a replica set. This makes the test
skip unless the default DB is actually standalone rather than
assuming it must be.

The other tests are wrapped in subtests to simplify their existing
skip behaviors.  Eventually, they need to be replaced with proper
tests that don't assume a hard-coded configuration.