Only raise wnote/jnote on special error strings
Generally, wnote and jnote should not be considered errors, but the
Bulk API QA test specifies two cases in which they must be.  Rather
than raising any wnote or jnote, this commit only raises an error if
specific wnote/jnote strings are detected.

These strings appear to be the same at least back to 1.8.5, so should
be safe to check.