Rename Bulk & WriteSelector to BulkWrite & BulkWriteView
I renamed Bulk to BulkWrite for greater clarity about what exactly is
being done in bulk and to make the next change sound better.

I renamed WriteSelector to BulkWriteView in anticipation of being
consistent with the fluent API which will have Collection and
CollectionView.  The idea is that a "FooView" constrains the Foo with a
document selection and only allows operations that are relevant after
find().