Switch MongoClient to use _URI to build connections
This replaces the long BUILD method with simple initialization based
on the MongoDB::_URI class.  The _Link class has changed to have all
parameters in the constructor, which simplifies usages.

The BUILD method loses support for replica sets, but this will be
replaced later when the cluster model and server selection are
implemented.