Revise socket handshake and report errors on selection failure
This commit revises the handshake algorithm to process the isMaster
result to update the topology before attempting to authenticate.

Errors though the process are recorded into the server descriptions
and used for more informative error messages if server selection
fails.  Specifically, on selection failure, the deployment members
are shown with type and error (if any).