clean up tag set logic and usage
This fixes the application of tag set lists to NEAREST, and brings
the names of variables and methods into alignment with the draft
server selection spec.

Also, the use of the term "tag set" and "tag set list" and related
variable names has been made more consistent with the spec.

The term "tags" is still used by the _Server class for consistency
with the ismaster command result and the SDAM spec.