implement insert_many and revise BulkWriteResult
BulkWriteResult now has the 'upserted' field again as an array
of documents as returned by the server.  The 'upserted_ids' field
is now a hash reference for consistency with the CRUD spec.
'upserted_ids' is generated lazily from 'upserted'.

The 'inserted' and 'inserted_ids' fields do the same thing for
insertions rather the upserts.

With these changes, 'insert_many' was implemented using the Bulk API
behind the scenes.