PERL-496 remove use_boolean global
The use_boolean global was documented as defaulting to false, but
was actually checked incorrectly and was always considered true.

Fixing this would change the default behavior.  While it could be fixed
with the opposite default, that pepetuates the use of global variables,
which are going away in the v1.x driver under development.

Therefore, this commit removes use_boolean and documents that it is
not supported.