add undocumented connected method for back compatibility
The v0.x series had an undocumented 'connected' method.  The semantics
of that made sense when one MongoClient object corresponded to one
socket, or to a collection of sockets, of which one was deemed the
master.

With v1.x, the being connected or not is largely irrelevant as the
driver reconnects as needed and just times out server selection if not
server is available to satisfy an operation.

This commit adds an undocumented method for compatibility that returns
true or false depending on whether a writable server is available.  Like
other such methods, it can block for up to the server selection timeout
trying to find a valid server.

While this may be unexpected, I think this is the most correct analogy
for what the old method did.