PERL-531 fix bulk update/replace validation for op_char
Bulk API never validated update/replace keys against anything
except '$'.  This commit changes that to delay the check until
execution time, after encoding.  It also only checks a single
key for efficiency, consistent with the CRUD API spec.

The update/remove document validation was factored into a new
role, MongoDB::Role::_UpdatePreEncoder, which is used by both the
regular _Update op as well as the _BulkWrite op.

This commits encodes the update/replace document *first* into a
BSON string and pulls the encoded first string back out to check it.

The test for whether a document is a legal replace/update document
checks the first key to see if it starts with '$' or not.  This
come *after* op_char replacement however.

It then blesses the string to MongoDB::BSON::Raw for inclusion in the
update command document.