PERL-127 encode small ints to Int32; encode BigInts to Int64
For ordinary Perl integers, they are encoded as BSON Int32 if they fit
and BSON Int64 otherwise.

For Math::BigInt objects, conversion now extracts the BigInt as a string
and process with strtoll.  The previous implementation violated
encapsulation and made assumptions about the inner structure of
Math::BigInt objects.  The result will then alwasy be encoded as BSON
Int64.