PERL-413 wire up max_time_ms configuration option
This commit adds 'max_time_ms' to Database, Collection and GridFS
classes and uses it as the default for any queries that support it
(but only on server 2.6+).

This is important because we'd rather get a database error than a socket
timeout, as the latter invalidates the connection and will trigger
topology scanning.

The max_time_ms.t tests have been expanded, but now only run with
a modern Test::Harness to ensure that they don't get run in parallel
with other tests and cause those tests to fail when failpoints are
turned on.

Queries take care to strip maxTimeMS for '$cmd*' collections, where
it is illegal.