PERL-546 remove 'client' and 'fetch' from DBRef
Providing a 'fetch' method is inconsistent with other MongoDB drivers,
which either never provided it, or have dropped it in the
next-generation drivers.

It also requires a 'client' attribute, which tightly couples BSON
decoding to the client model.

In the current BSON codec implementation, it leads to a circular
reference.  When weakened, it leads to memory bugs under threads.

Therefore, in order to decouple the BSON data model from the
driver, restore consistency with other drivers, and avoid
circularity/memory problems, this commit removes the 'fetch' feature
completely.