PERL-553 prevent duplicate _id inserting IxHash and array docs
This commit fixes a bug inserting a Tie::IxHash or array reference
document with an existing _id field where the _id field would exist
twice in the BSON document sent to the server.

This is nearly impossible to test from within the Perl driver, but I
manually verified it using the mongo shell.