Moo-ify internal classes for insertion
This commit converts the classes used in insertion to Moo and switches
off type assertions unless $ENV{PERL_MONGO_ENABLE_ASSERTS} is true.
Various coercions and native traits are re-implemented by hand.

This covers everything for insert_one.  If it prove effective,
it needs to be extended to the rest of the internal classes.