convert more internal classes to private constructor
This commit removes attribute defaults from results
classes (and the _EncodedDoc class) and replaces them
with a 'required' modifier.  All attributes are now
passed in the constructor call.

Classes are enhanced with the MongoDB::_PrivateConstructor
role, which uses '_new' to bless arguments directly into
the classs, bypassing assertions for speed.