Split up long sentence for readability.
Long sentences get hard to read (at least for me), because you have to
hold the grammatical state of the sentence in your head as you read.
Just adding a period can allow the reader to "write to disk", as it
were, and be able to absorb the next piece of information better.