Revise storage of cursor IDs
Previously, cursor IDs were always stored as opaque 8-byte strings.
This meant that for command cursors, which are decoded from a BSON
document and give cursor IDs as perl integers (or Math::BigInt objects),
cursor ID needed to be packed back into the 8-byte form.

In preparation for the change to the 'getMore' command instead of
OP_GET_MORE in MongoDB 3.2, this commit stores the cursor ID in one
of three forms:

* A MongoDB::_CursorID object, which is a blessed string reference to
  the 8-byte cursorID.  This form comes from parsing OP_REPLY if the
  cursor ID is non-zero.
* A Perl integer, which comes from OP_REPLY if the cursor ID is zero (as 0)
  or from a command that returns a cursor.
* A Math::BigInt object, which comes from a command that returns a cursor
  if the cursor ID requires more than 32 bits and Perl only supports 32 bit
  integers.

As a result, the MongoDB::_Protocol class now handles blessing to
MongoDB::_CursorID when appropriate and converting from any of the three
forms to the packed Int64 form for the wire protocol for OP_GET_MORE
and OP_KILL_CURSORS.

This will ensure the legacy operations work as expected, but on 3.2, will
allow the cursor ID returned from the find, aggregate, getMore, etc.
commands to be used as is in subsequent getMore commands.