PERL-567 Fix tests for a replica set on 27017
If a replica set member is running on port 27017 and there is no
$ENV{MONGOD} set, then the tests were making a direct connection
to 27017 instead of discovering the replica set.  This broke
a test that needed a secondary.

This commit has the "skip without mongodb" test helper set $ENV{MONGOD}
with the replica set name if this configuration is detected.

Additional orchestration tool changes were made to test this fix.