PERL-597 Check findAndModify for writeConcernErrors
When the writeConcern on findAndModify fails, the resulting document still
comes back with ok: 1.  For most write commands, the write result object is
used to check for such errors. But since findAndModify based commands don't
have result objects (since they return documents), we must check the
command result directly for a writeConcernError field. If it is present,
then writeConcern failed and we throw the appropriate exception.