Consolidate _Query and Op::_Query
The MongodB::_Query class was merely a deferred MongoDB::Op::_Query factory
with different attribute names.  This redundancy means an extra object is
created for all queries, even though many aren't subsequently modified.

This commit consolidates the two and allows MongoDB::Cursor to directly
modify the Op::_Query.  This required that Op::_Query parameters become
mutable, but as this is a private class, that shouldn't be a significant
source of bugs.