Factor common operation attributes into separate roles
This commit move db_name into Role::_DatabaseOp and adds a new
Role::_CollectionOp for coll_name and full_name.

It also changes some use of 'ns' to 'full_name' for consistency and to be
able to use the new role everywhere that it makes sense to do so.

Some use of operations needed to add 'full_name', as it is required, even
though it can be a dummy field in some cases.  Trying to factor that out
separately adds complexity that probably obscures more than it clarifies.
And we don't want to get rid of full_name entirely because legacy
operations need it and we want to generate it once early and pass it along
rather then recompute for every operation.