PERL-677 Make fsync tests more robust
This removes tests for specific error strings/fields.  It also
removes the 'async' test as this test is really only testing that
we pass through arguments to the server, which is redundant with the
lock test.  We also add a confirmation test that the unlock command
actually unlocked the database.